Divide App component from index
한 파일이 한 가지 역할만 하도록 수정하였습니다.
핸들러 함수를 넘겨줄때의 명칭은 아직 어떻게 해야할지 갈피를 못잡겠네요.
